+++ apps/meteor/tests/e2e/page-objects/fragments/sidebar.ts
@@ -79,7 +79,7 @@ export class Sidebar {
 		await item.hover();
 		await item.focus();
 		await item.locator('.rcx-sidebar-item__menu').click();
-		await this.page.getByRole('option', { name: 'Mark Unread' }).click();
+		await this.page.getByRole('menuitem', { name: 'Mark Unread' }).click();
 	}
 
 	getCollapseGroupByName(name: string): Locator {