+++ apps/meteor/tests/end-to-end/api/rooms.ts
@@ -2,7 +2,9 @@ import fs from 'fs';
 import path from 'path';
 
 import type { Credentials } from '@rocket.chat/api-client';
-import type { IMessage, IRoom, ITeam, IUpload, IUser, ImageAttachmentProps, SettingValue } from '@rocket.chat/core-typings';
+import type { IMessage, IRole, IRoom, ITeam, IUpload, IUser, ImageAttachmentProps, SettingValue } from '@rocket.chat/core-typings';
+import { TEAM_TYPE } from '@rocket.chat/core-typings';
+import { Random } from '@rocket.chat/random';
 import { assert, expect } from 'chai';
 import { after, afterEach, before, beforeEach, describe, it } from 'mocha';
 
@@ -11,8 +13,9 @@ import { getCredentials, api, request, credentials } from '../../data/api-data';
 import { sendSimpleMessage, deleteMessage } from '../../data/chat.helper';
 import { imgURL } from '../../data/interactions';
 import { getSettingValueById, updateEEPermission, updatePermission, updateSetting } from '../../data/permissions.helper';
+import { assignRoleToUser, createCustomRole, deleteCustomRole } from '../../data/roles.helper';
 import { createRoom, deleteRoom } from '../../data/rooms.helper';
-import { deleteTeam } from '../../data/teams.helper';
+import { createTeam, deleteTeam } from '../../data/teams.helper';
 import { password } from '../../data/user';
 import type { TestUser } from '../../data/users.helper';
 import { createUser, deleteUser, login } from '../../data/users.helper';
@@ -3631,6 +3634,533 @@ describe('[Rooms]', () => {
 				.expect((res: Response) => {
 					expect(res.body).to.have.property('success', false);
 				});
+		});
+	});
+
+	describe('[/rooms.membersOrderedByRole]', () => {
+		const isEnterprise = Boolean(process.env.IS_EE);
+
+		let testChannel: IRoom;
+		let ownerUser: IUser;
+		let moderatorUser: IUser;
+		let memberUser1: IUser;
+		let memberUser2: IUser;
+		let customRole: IRole;
+
+		let ownerCredentials: { 'X-Auth-Token': string; 'X-User-Id': string };
+
+		before(async () => {
+			[ownerUser, moderatorUser, memberUser1, memberUser2] = await Promise.all([
+				createUser({ username: `a_${Random.id()}`, roles: ['admin'] }),
+				createUser({ username: `b_${Random.id()}` }),
+				createUser({ username: `c_${Random.id()}` }),
+				createUser({ username: `d_${Random.id()}` }),
+			]);
+
+			ownerCredentials = await login(ownerUser.username, password);
+
+			customRole = await createCustomRole({
+				name: `customRole.${Random.id()}`,
+				scope: 'Subscriptions',
+				description: 'Custom Role',
+			});
+
+			// Create a public channel
+			const roomCreationResponse = await createRoom({
+				type: 'c',
+				name: `rooms.membersOrderedByRole.test.${Date.now()}`,
+				credentials: ownerCredentials,
+			});
+			testChannel = roomCreationResponse.body.channel;
+
+			await Promise.all(
+				[moderatorUser._id, memberUser1._id, memberUser2._id].map((userId) =>
+					request
+						.post(api('channels.invite'))
+						.set(ownerCredentials)
+						.send({
+							roomId: testChannel._id,
+							userId,
+						})
+						.expect(200),
+				),
+			);
+
+			await request
+				.post(api('channels.addModerator'))
+				.set(ownerCredentials)
+				.send({
+					roomId: testChannel._id,
+					userId: moderatorUser._id,
+				})
+				.expect(200);
+		});
+
+		after(async () => {
+			await deleteRoom({ type: 'c', roomId: testChannel._id });
+			await Promise.all([ownerUser, moderatorUser, memberUser1, memberUser2].map((user) => deleteUser(user)));
+			if (isEnterprise && customRole) {
+				await deleteCustomRole({ roleId: customRole._id });
+			}
+		});
+
+		it('should return a list of members ordered by owner, moderator, then members by default', async () => {
+			const response = await request
+				.get(api('rooms.membersOrderedByRole'))
+				.set(credentials)
+				.query({
+					roomId: testChannel._id,
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			expect(response.body.members).to.be.an('array');
+
+			const [first, second, ...rest] = response.body.members;
+			expect(first.username).to.equal(ownerUser.username);
+			expect(second.username).to.equal(moderatorUser.username);
+
+			const memberUsernames = rest.map((m: any) => m.username);
+			expect(memberUsernames).to.include(memberUser1.username);
+			expect(memberUsernames).to.include(memberUser2.username);
+
+			expect(response.body).to.have.property('total');
+			expect(response.body.total).to.be.gte(4);
+		});
+
+		it('should support sorting by role in descending priority', async () => {
+			const response = await request
+				.get(api('rooms.membersOrderedByRole'))
+				.set(credentials)
+				.query({
+					roomId: testChannel._id,
+					sort: '{"rolePriority":-1}',
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			const [first, second, third, fourth] = response.body.members;
+
+			expect(first.username).to.equal(memberUser1.username);
+			expect(second.username).to.equal(memberUser2.username);
+			expect(third.username).to.equal(moderatorUser.username);
+			expect(fourth.username).to.equal(ownerUser.username);
+		});
+
+		it('should support pagination', async () => {
+			const response = await request
+				.get(api('rooms.membersOrderedByRole'))
+				.set(credentials)
+				.query({
+					roomId: testChannel._id,
+					count: 2,
+					offset: 0,
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			expect(response.body.members).to.have.lengthOf(2);
+			expect(response.body.total).to.be.gte(4);
+		});
+
+		it('should return matched members when using filter param', async () => {
+			const response = await request
+				.get(api(`rooms.membersOrderedByRole`))
+				.set(credentials)
+				.query({
+					roomId: testChannel._id,
+					filter: memberUser1.username,
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			expect(response.body.members).to.have.lengthOf(1);
+			expect(response.body.members[0]).have.property('username', memberUser1.username);
+		});
+
+		it('should return empty list if no matches (e.g., filter by status that no one has)', async () => {
+			const response = await request
+				.get(api(`rooms.membersOrderedByRole`))
+				.set(credentials)
+				.query({
+					'roomId': testChannel._id,
+					'status[]': 'SomeRandomStatus',
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			expect(response.body.members).to.be.an.empty('array');
+		});
+
+		it('should support custom sorting by username descending', async () => {
+			const response = await request
+				.get(api('rooms.membersOrderedByRole'))
+				.set(credentials)
+				.query({
+					roomId: testChannel._id,
+					sort: JSON.stringify({ username: -1 }),
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			const usernames = response.body.members.map((m: any) => m.username);
+
+			const expected = [
+				ownerUser.username, // since owner
+				moderatorUser.username, // since moderator
+				memberUser2.username,
+				memberUser1.username,
+			];
+
+			expect(usernames).to.deep.equal(expected);
+		});
+
+		it('should not be affected by custom roles when sorting', async () => {
+			if (!isEnterprise) {
+				return;
+			}
+			await Promise.all([
+				assignRoleToUser({ username: moderatorUser.username as string, roleId: customRole._id }),
+				assignRoleToUser({ username: memberUser2.username as string, roleId: customRole._id }),
+			]);
+
+			const response = await request
+				.get(api('rooms.membersOrderedByRole'))
+				.set(credentials)
+				.query({
+					roomId: testChannel._id,
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200);
+
+			expect(response.body).to.have.property('success', true);
+			const [first, second, third, fourth] = response.body.members;
+
+			expect(first.username).to.equal(ownerUser.username);
+			expect(second.username).to.equal(moderatorUser.username);
+			expect(third.username).to.equal(memberUser1.username);
+			expect(fourth.username).to.equal(memberUser2.username);
+		});
+
+		describe('Additional Visibility Tests', () => {
+			let outsiderUser: IUser;
+			let insideUser: IUser;
+			let nonTeamUser: IUser;
+			let outsiderCredentials: { 'X-Auth-Token': string; 'X-User-Id': string };
+			let insideCredentials: { 'X-Auth-Token': string; 'X-User-Id': string };
+			let nonTeamCredentials: { 'X-Auth-Token': string; 'X-User-Id': string };
+
+			let privateChannel: IRoom;
+			let publicChannel: IRoom;
+			let publicTeam: ITeam;
+			let privateTeam: ITeam;
+			let privateChannelInPublicTeam: IRoom;
+			let publicChannelInPublicTeam: IRoom;
+			let privateChannelInPrivateTeam: IRoom;
+			let publicChannelInPrivateTeam: IRoom;
+
+			before(async () => {
+				[outsiderUser, insideUser, nonTeamUser] = await Promise.all([
+					createUser({ username: `e_${Random.id()}` }),
+					createUser({ username: `f_${Random.id()}` }),
+					createUser({ username: `g_${Random.id()}` }),
+				]);
+				[outsiderCredentials, insideCredentials, nonTeamCredentials] = await Promise.all([
+					login(outsiderUser.username, password),
+					login(insideUser.username, password),
+					login(nonTeamUser.username, password),
+				]);
+
+				// Create a public team and a private team
+				[publicTeam, privateTeam] = await Promise.all([
+					createTeam(insideCredentials, `rooms.membersOrderedByRole.team.public.${Random.id()}`, TEAM_TYPE.PUBLIC, [
+						outsiderUser.username as string,
+					]),
+					createTeam(insideCredentials, `rooms.membersOrderedByRole.team.private.${Random.id()}`, TEAM_TYPE.PRIVATE, [
+						outsiderUser.username as string,
+					]),
+				]);
+
+				const [
+					privateInPublicResponse,
+					publicInPublicResponse,
+					privateInPrivateResponse,
+					publicInPrivateResponse,
+					privateRoomResponse,
+					publicRoomResponse,
+				] = await Promise.all([
+					createRoom({
+						type: 'p',
+						name: `teamPublic.privateChannel.${Date.now()}`,
+						credentials: insideCredentials,
+						extraData: {
+							teamId: publicTeam._id,
+						},
+					}),
+					createRoom({
+						type: 'c',
+						name: `teamPublic.publicChannel.${Date.now()}`,
+						credentials: insideCredentials,
+						extraData: {
+							teamId: publicTeam._id,
+						},
+					}),
+					createRoom({
+						type: 'p',
+						name: `teamPrivate.privateChannel.${Date.now()}`,
+						credentials: insideCredentials,
+						extraData: {
+							teamId: privateTeam._id,
+						},
+					}),
+					createRoom({
+						type: 'c',
+						name: `teamPrivate.publicChannel.${Date.now()}`,
+						credentials: insideCredentials,
+						extraData: {
+							teamId: privateTeam._id,
+						},
+					}),
+					createRoom({
+						type: 'p',
+						name: `rooms.membersOrderedByRole.private.${Date.now()}`,
+						credentials: insideCredentials,
+					}),
+					createRoom({
+						type: 'c',
+						name: `rooms.membersOrderedByRole.public.${Date.now()}`,
+						credentials: insideCredentials,
+					}),
+				]);
+
+				privateChannelInPublicTeam = privateInPublicResponse.body.group;
+				publicChannelInPublicTeam = publicInPublicResponse.body.channel;
+				privateChannelInPrivateTeam = privateInPrivateResponse.body.group;
+				publicChannelInPrivateTeam = publicInPrivateResponse.body.channel;
+				privateChannel = privateRoomResponse.body.group;
+				publicChannel = publicRoomResponse.body.channel;
+			});
+
+			after(async () => {
+				await Promise.all([
+					deleteRoom({ type: 'p', roomId: privateChannel._id }),
+					deleteRoom({ type: 'c', roomId: publicChannel._id }),
+					deleteRoom({ type: 'p', roomId: privateChannelInPublicTeam._id }),
+					deleteRoom({ type: 'c', roomId: publicChannelInPublicTeam._id }),
+					deleteRoom({ type: 'p', roomId: privateChannelInPrivateTeam._id }),
+					deleteRoom({ type: 'c', roomId: publicChannelInPrivateTeam._id }),
+				]);
+
+				await Promise.all([deleteTeam(credentials, publicTeam.name), deleteTeam(credentials, privateTeam.name)]);
+
+				await Promise.all([deleteUser(outsiderUser), deleteUser(insideUser), deleteUser(nonTeamUser)]);
+			});
+
+			it('should not fetch private room members by user not part of room', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(outsiderCredentials)
+					.query({ roomId: privateChannel._id })
+					.expect('Content-Type', 'application/json')
+					.expect(404)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+					});
+			});
+
+			it('should fetch private room members by user who is part of the room', async () => {
+				const response = await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(insideCredentials)
+					.query({ roomId: privateChannel._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200);
+
+				expect(response.body.success).to.be.true;
+				expect(response.body.members).to.be.an('array');
+			});
+
+			it('should fetch public room members by user who is part of the room', async () => {
+				const response = await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(insideCredentials)
+					.query({ roomId: publicChannel._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200);
+
+				expect(response.body.success).to.be.true;
+				expect(response.body.members).to.be.an('array');
+			});
+
+			it('should fetch public room members by user not part of room - because public', async () => {
+				await updatePermission('view-c-room', ['admin', 'user', 'guest']);
+				const response = await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(outsiderCredentials)
+					.query({ roomId: publicChannel._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200);
+
+				expect(response.body.success).to.be.true;
+				expect(response.body.members).to.be.an('array');
+			});
+
+			it('should fetch a private channel members inside a public team by someone part of the room ', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(insideCredentials)
+					.query({ roomId: privateChannelInPublicTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should not fetch a private channel members inside a public team by someone not part of the room, but part of team', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(outsiderCredentials)
+					.query({ roomId: privateChannelInPublicTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(404)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+					});
+			});
+
+			it('should not fetch a private channel members inside a public team by someone not part of the team ', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(nonTeamCredentials)
+					.query({ roomId: privateChannelInPublicTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(404)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+					});
+			});
+
+			it('should fetch a public channel members inside a public team by someone part of the room ', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(insideCredentials)
+					.query({ roomId: publicChannelInPublicTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should fetch a public channel members inside a public team by someone not part of the room, but part of team', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(outsiderCredentials)
+					.query({ roomId: publicChannelInPublicTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should fetch a public channel members inside a public team by someone not part of the team ', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(nonTeamCredentials)
+					.query({ roomId: publicChannelInPublicTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should fetch a public channel members inside a private team by someone part of the room', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(insideCredentials)
+					.query({ roomId: publicChannelInPrivateTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should fetch a public channel members inside a private team by someone not part of the room, but part of team', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(outsiderCredentials)
+					.query({ roomId: publicChannelInPrivateTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should not fetch a public channel members inside a private team by someone not part of team', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(nonTeamCredentials)
+					.query({ roomId: publicChannelInPrivateTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(404)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+					});
+			});
+
+			it('should fetch a private channel members inside a private team by someone part of the room', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(insideCredentials)
+					.query({ roomId: privateChannelInPrivateTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(200)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', true);
+						expect(res.body.members).to.be.an('array');
+					});
+			});
+
+			it('should not fetch a private channel members inside a private team by someone not part of the room, but part of team', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(outsiderCredentials)
+					.query({ roomId: privateChannelInPrivateTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(404)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+					});
+			});
+
+			it('should not fetch a private channel members inside a private team by someone not part of team', async () => {
+				await request
+					.get(api('rooms.membersOrderedByRole'))
+					.set(nonTeamCredentials)
+					.query({ roomId: privateChannelInPrivateTeam._id })
+					.expect('Content-Type', 'application/json')
+					.expect(404)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+					});
+			});
 		});
 	});
 });